% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_LandParams.R
\docType{class}
\name{ArtificialLandscape}
\alias{ArtificialLandscape}
\title{Create an Artificial Landscape}
\usage{
ArtificialLandscape(propSuit = 0.5, K_or_DensDep = 10, Resolution = 100, dimX = 65, dimY = 65,
                    fractal = FALSE, hurst, continuous = FALSE, minPct, maxPct)
}
\arguments{
\item{propSuit}{Proportion of suitable habitat cells, defaults to \eqn{0.5}.}

\item{K_or_DensDep}{determines the demographic density dependence of the modelled species and is given in units of the number of individuals per hectare (defaults to \eqn{10}).
Its precise interpretation depends on the structure of the species' \code{\link[RangeShiftR]{Demography}}:\cr
If combined with a \code{\link[RangeShiftR]{StageStructure}}d model, \code{K_or_DensDep} will be used as the \emph{strength of demographic density dependence} \ifelse{html}{\out{b<sup>-1</sup>}}{\eqn{1/b}}.
If combined with a non-structured model, \code{K_or_DensDep} will be interpreted as \emph{limiting carrying capacity} \eqn{K}.\cr}

\item{Resolution}{Cell size in meters, defaults to \eqn{100}. (integer)}

\item{dimX, dimY}{Number of cells along the x- and y-axis, both default to \eqn{65}. (integer) \cr
If \code{fractal=TRUE}, \code{dimX} and \code{dimY} must be powers of \eqn{2} plus \eqn{1} (\eqn{2^n+1}) and \code{dimX} \eqn{\le} \code{dimY} .}

\item{fractal}{If \code{FALSE} (default), a random landscape is generated. Each cell has a certain probability of being a habitat.\cr
If \code{TRUE}, a fractal landscape is generated. These landscapes contain a greater structure than random landscapes but less than a completely deterministic one.}

\item{hurst}{Required if \code{fractal=TRUE}: Hurst exponent. Can be any number in the open interval \eqn{(0,1)}.}

\item{continuous}{Use continuous or binary habitat values to describe each cell?\cr If \code{FALSE} (default), the resulting landscape
is binary, with the two values \emph{Habitat} and \emph{Matrix} (i.e. Non-habitat).\cr If \code{TRUE}, each cell is given a continuous value
between \eqn{0} and \eqn{1}, which describes the percentage of habitat cover or habitat quality.
within a cell. The effective \code{K_or_DensDep} of that cell is calculated as the respective fraction of the value of \eqn{K_or_DensDep}.}

\item{minPct, maxPct}{Required if \code{continuous=TRUE}: Minimum and Maximum percentage of habitat cover within a cell. Can be any number in the open interval \eqn{(0,1)},
\code{maxPct} may be exactly \eqn{1}. \code{minPct} must be smaller than \code{maxPct}.}
}
\value{
A parameter object of class "ArtificialLandscape"
}
\description{
An artificial landscape can be generated with a random or fractal spatial structure, and may be using binary or continuous habitat values to characterise each cell.
}
\details{
For theoretical studies which might be related to fundamental questions in eco-evolutionary dynamics or strategic questions concerning
conservation ecology, it is often desirable to use artificial landscapes.\cr
Fractal landscapes are characterised by possessing greater structure than a completely random landscape, but less than a
completely deterministic one \insertCite{with1997application}{RangeShiftR} – but note that the spatial structure of landscapes fragmented
by human activities is often not fractal in nature and, depending upon the research question, other landscape generators may be more
appropriate \insertCite{@see review by @pe2013simple}{RangeShiftR}.

Internally generated artificial landscapes may not be dynamic.

The fractal landscape generator implemented in \emph{RangeShiftR} uses the midpoint displacemeant algorithm \insertCite{saupe1988algorithms}{RangeShiftR}.
The Hurst exponent, often referred to as \eqn{H}, describes the degree of spatial autocorrelation of the landscape configuration.
Values close to \eqn{0} represent a low autocorrelation but the generated landscapes still aren't completely spatially independent.
On the contrary, values close to \eqn{1} represent high autocorrelation, i.e. high habitat aggregation.\cr

Note that more complex algorithms are available for providing fractals where setting \eqn{H = 0.0} results in no spatial autocorrelation
\insertCite{@see @chipperfield2011updated}{RangeShiftR}. For applications where the embedded algorithm is not sufficient, we recommend
to import landscapes generated by these alternative algorithms.
}
\references{
\insertAllCited{}
}
\author{
Anne-Kathleen Malchow
}
