% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_handling.R
\name{ColonisationStats}
\alias{ColonisationStats}
\title{ColonisationStats}
\usage{
ColonisationStats(x, y = getwd(), years = numeric(0), maps = FALSE)
ColonisationStats(x, y = NULL,    years = numeric(0))
}
\arguments{
\item{x, y}{Either the parameter master (\code{x}) and RS directory path (\code{y}) of the simulation or, alternatively,\cr
the population output as a dataframe (\code{x}). In this case \code{y} is an optional parameter taking the patch map(s) as a raster layer (stack),
which will then be used to produce maps of the patch statistics. For more info on this, see the Details.}

\item{years}{Years at which the probabilty of occupancy over all replicates will be calculated.}

\item{maps}{Only in the parameter master (=\code{x}) notation: For each given \code{year}, uses the current patch map (in case of a dynamic landscape, otherwise the static patch map)
to produce a raster of occupancy probabilties. For time to colonisation, uses the current patch map of last year of population record.}
}
\value{
a list with the elements
\code{occ_prob} (a dataframe with mean (over all replicates) probabilty of occupancy for each patch (in rows) at the given year(s) in columns),\cr
\code{col_time} (a dataframe with first recorded year of colonisation for each patch; with replicates in columns and rows named by patch-IDs),\cr
Optional:\cr\cr
\code{patch_occ_prob} (a raster (stack) with the data from \code{occ_prob} stored in the respective patch cells per given year),\cr
\code{patch_col_time} (a raster with the data from \code{col_time_mean} stored in their respective patch cells)
}
\description{
This function produces patch statistics and maps on occupancy probability and mean time to colonisation.
}
\details{
It uses the RangeShiftR 'population' output data.

In the population dataframe (=\code{x}) notation, there are 3 options on how many maps to produce:
(1) y = NULL: no raster maps produced.
(2) y is a RasterLayer: All statistics projected onto the same patch map.
(3) y is a RasterStack with 2 layers: The first will be used for probabilty of occupancy for all years, the second will be used for time to colonisation.
(4) y is a RasterStack with length(years)+1 layers: The first ones will be used for probabilty of occupancy on each year, the last for time to colonisation.
}
