% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DispersalParams.R
\docType{class}
\name{Dispersal}
\alias{Dispersal}
\title{Set Dispersal Parameters}
\usage{
Dispersal(Emigration = Emigration(),
          Transfer   = DispersalKernel(),
          Settlement = Settlement())
}
\arguments{
\item{Emigration}{The first phase of dispersal; determines if an individual leaves its natal patch.}

\item{Transfer}{(or transience) The second phase of dispersal; consists of the movement of an individual departing from its natal patch towards
a potential new patch, ending with settlement or mortality. This movement can be modelled by one of three alternative processes:\cr
- Dispersal kernel: use \code{\link[RangeShiftR]{DispersalKernel}}\cr
- Stochastic movement simulator (SMS): use \code{\link[RangeShiftR]{SMS}}\cr
- Correlated random walk (CRW): use \code{\link[RangeShiftR]{CorrRW}}}

\item{Settlement}{(or immigration) The last phase of dispersal; determines when the individual stops in a new cell or patch of
breeding habitat.}
}
\value{
a parameter object of class "DispersalParams"
}
\description{
Dispersal is defined as movement leading to spatial gene flow. It typically involves three phases, which are all modelled explicitly:
\code{\link[RangeShiftR]{Emigration}}, \code{\link[RangeShiftR]{Transfer}} and \code{\link[RangeShiftR]{Settlement}}.\cr
The specific parameters of each phase are set through their respective functions. For more details, see their documentation.\cr
\cr
The potential evolution of several key dispersal traits is implemented through the possibility of inter-individual variability and heritability.
This option (called \code{IndVar}) can be enabled for each dispersal phase in their respective modules. See the Details below for information on how
to set the associated parameters. Additionally, the \code{\link[RangeShiftR]{Genetics}} module must be defined.
}
\details{
Dispersal is defined as movement leading to spatial gene flow, and it typically involves three phases:
emigration, transfer and settlement
 \insertCite{stenseth1992,clobert2001,clobert2009,clobert2012,bowler2005,ronce2007}{RangeShiftR}.\cr

The key role of dispersal in species persistence and responses to environmental change is increasingly recognized \insertCite{travis2013dispersal}{RangeShiftR}.
Moreover, the importance of modelling dispersal as a complex process, explicitly considering its three phases, each of which has
´its own mechanisms and costs, has been recently highlighted \insertCite{bonte2012costs,travis2012modelling,travis2013dispersal}{RangeShiftR}.
The implementation of the dispersal process in \emph{RangeShiftR} is based on these recent frameworks and the substantial dispersal
theory that has been developed so far \insertCite{clobert2012}{RangeShiftR}.

It is possible to model all three phases - emigration, transfer and settlement - with stage- and/or sex-specific parameters via the
switches \code{StageDep} and \code{SexDep} in the respective functions. In the case of sex-dependence, the number of traits is doubled, with
one set coding for the trait in females and the other for the trait in males. As well as being sex-biased, all dispersal phases can be
stage-biased, meaning that parameters can vary for different stage or age classes.

The options of inter-individual variability in the various dispersal traits can be enabled by setting \code{IndVar=TRUE} for the respective dispersal phase
module and defining the genetic module to simulate heritability and evolution of traits (\code{\link[RangeShiftR]{Genetics}}).\cr
For each such heritable dispersal trait, this requires to set the mean and standard deviation of the distribution of trait values
in the initial population (instead of a constant value each, as in the case of \code{IndVar=FALSE}),
as well as the \code{TraitScaleFactor} for each heritable trait (see \code{\link[RangeShiftR]{Genetics}} documentaion).\cr

Note, however, that not all combinations of sex-/stage-dependencies with inter-individual variability are implemented.\cr
}
\references{
\insertAllCited{}
}
\author{
Anne-Kathleen Malchow
}
