% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_LandParams.R
\docType{class}
\name{ImportedLandscape}
\alias{ImportedLandscape}
\title{Import a Landscape from file}
\usage{
ImportedLandscape(LandscapeFile, Resolution = 100, HabPercent = FALSE,
                  Nhabitats, K_or_DensDep = 10,
                  PatchFile = "NULL",
                  CostsFile = "NULL",
                  DynamicLandYears = 0,
                  SpDistFile = "NULL", SpDistResolution)
}
\arguments{
\item{LandscapeFile}{Filename(s) of the landscape habitat map(s) which shall be imported from the Inputs-folder. See the Details for information on the required format.}

\item{Resolution}{Cell size in meters, defaults to \eqn{100}. (integer)}

\item{HabPercent}{If \code{FALSE} (default), unique integer habitat codes are expected in the imported map to characterise the habitat of each cell. This requires to set \code{Nhabitats}. \cr
If \code{TRUE}, continuous values are expected, ranging from \eqn{0.0} to \eqn{100.0}, that represent percentages of habitat cover or quality.\cr
Make sure your imported landscape file uses the specified standard (see Details below).}

\item{Nhabitats}{Required if \code{HabPercent=FALSE}: Number of habitats in the imported landscape if unique integer habitat codes are used. (integer)}

\item{K_or_DensDep}{determines the demographic density dependence of the modelled species and is given in units of the number of individuals per hectare (defaults to \eqn{10}).
Its precise interpretation depends on the structure of the species' \code{\link[RangeShiftR]{Demography}}:\cr
If combined with a \code{\link[RangeShiftR]{StageStructure}}d model, \code{K_or_DensDep} will be used as the \emph{strength of demographic density dependence} \ifelse{html}{\out{b<sup>-1</sup>}}{\eqn{1/b}}.
If combined with a non-structured model, \code{K_or_DensDep} will be interpreted as \emph{limiting carrying capacity} \eqn{K}.\cr
The expected format:\cr
If \code{HabPercent=FALSE}: a vector of length \code{Nhabitats}, specifying the respective \code{K_or_DensDep} for every habitat code.\cr
If \code{HabPercent=TRUE}: \code{K_or_DensDep} is interpreted as the maximum \code{K_or_DensDep} reached in cells with \eqn{100}\% habitat. All other cells hold the respective fraction of \eqn{K_or_DensDep}.}

\item{PatchFile}{Filename(s) of the patch map(s) which shall be imported, default is \code{NULL}.}

\item{CostsFile}{Filename(s) of the SMS cost map(s) which shall be imported, default is \code{NULL}.}

\item{DynamicLandYears}{Integer vector indicating the years of landscape changes. For a non-dynamic landscape its only entry is \eqn{0} (default).
For a dynamic landscape, \code{DynamicLandYears} lists the years in which the corresponding habitat maps in \code{LandscapeFile} and - if applicable - their respective patch and/or costs
maps (in \code{PatchFile},\code{CostsFile}) are loaded and used in the simulation. More details below.}

\item{SpDistFile}{Filename of the species initial distribution map which shall be imported (*.txt). Default is \code{NULL}.}

\item{SpDistResolution}{Required if \code{SpDistFile} is given: Cell size of the distribution map in meters. (integer) Must be an integer multiple of the landscape resolution.}
}
\value{
A parameter object of class ImportedLandscape
}
\description{
Provide the filename(s) (\code{LandscapeFile}) of the map(s) to be imported, their
resolution and, if applicable, the number of habitat codes (\code{Nhabitats})
as well as their respective demographic density dependence (\code{K_or_DensDep}).

For a dynamic landscape, the year in which each landscape is loaded has to be provided.

Other, optional input maps are:\cr
- Patch map(s) to define habitat patches,\cr
- SMS cost map(s) to define landscape resistance to,\cr
- a distribution map to define an initial species distribution.
}
\details{
\emph{RangeShiftR} requires every input map to be a text file in ESRI ASCII raster format, which has the following six header lines:
\tabular{ll}{\code{ncols} \tab Number of columns \cr
\code{nrows} \tab Number of rows \cr
\code{xllcorner} \tab x-coordinate (longitude) of the lower-left corner \cr
\code{yllcorner} \tab y-coordinate (latitude) of the lower-left corner \cr
\code{cellsize} \tab Resolution (in meters) \cr
\code{NODATA_value} \tab Value for cells having missing data (usually -9999) }

The rest of the file is a grid containing a value for each cell, one line per row. \code{RangeShiftR} can read-in two different types of habitat maps:\cr
\itemize{
    \item \emph{Raster with habitat codes} (\code{HabPercent=FALSE})\cr In this option each habitat or land-cover type has a unique integer code. Each cell in the file contains a single habitat code and \eqn{100} percent coverage is assumed for the cell. The landscape is therefore composed of discrete habitat cells. The codes are required to be sequential integers starting from \eqn{1} and ranging to \code{Nhabitats}.\cr
    \item \emph{Raster with habitat quality} (\code{HabPercent=TRUE})\cr Each cell in the landscape is assigned a continuous percentage value between \eqn{0.0} and \eqn{100.0} of the maximum \code{K_or_DensDep}. There are no explicit habitat or land-cover types. This allows integrating different methods for calculating the habitat suitability for a given species. For example, qualities can result from
 different methods of suitability modelling, which incorporate multiple variables like habitat types, elevation, climate, etc. In the current version of the program, a straight-line relationship between the habitat quality and the actual density-dependence of the population dynamics (i.e. \eqn{K} or \eqn{1/b} in a non-structured or structured population, respectively) is assumed.
 Therefore, the quality should be scaled accordingly in case of a curvilinear relationship.
 }

\emph{Patch map} \cr
The simulation can be run as a \emph{patch-based model} on the same habitat map described above. An additional file must be provided through \code{PatchFile}: a raster map of the same landscape, where
each cell contains the ID number of the patch to which it belongs. Each patch must have a unique positive integer ID. The ID of every cell that does not belong to a patch (i.e. non-habitat/matrix) must be zero.
Note that a single patch is the unit at which the density dependence in the population dynamics acts. Therefore, a patch can be discontinuous, i.e. it can contain cells that do not belong to the patch if they
are assumed not to affect the dynamics, or on the other hand, patch cells that are not physically contiguous to the rest of the patch cells.

\emph{Costs layer} \cr
Only used if the simulation is run with \code{\link[RangeShiftR]{SMS}} as its transfer module and the landscapes resistance to movement is given via a costs raster map (see argument \code{Costs} in \code{SMS()}).
In this case, the specified map has to match the landscape raster in extent, coordinates and resolution, and each cell contains a cost value, with the minimal possible cost being \eqn{1}.
Importing a cost layer is the only option when the landscape comprises habitat coverage or quality.

\emph{Initial distribution} \cr
A \emph{species distribution map} can be overlaid on top of the habitat map and can be used to define an initial distribution. The map is provided through \code{SpDistFile} must be in raster format and be aligned with the landscape map, i.e. the coordinates of the lower-left corner must be the same. The extent of the map does not have to be necessarily
the same as the landscape. The resolution can be the same or coarser, provided that it is a multiple of the landscape resolution. For example, if the landscape cell size is \eqn{250m}, the species distribution can be at the resolution of \eqn{250m}, \eqn{500m}, \eqn{750m}, \eqn{1000m} etc.
Each cell of the species distribution map must contain either \eqn{0} (species absent or not recorded) or \eqn{1} (species present).

\emph{Dynamic landscapes} \cr
An imported landscape may be dynamic, i.e. the attributes of cells (either habitat class or quality index) and its patch number (if the model is patch-based) may be changed at specified years during the course of
a simulation. Note that any landscape change occurs at the start of the year, i.e. before the first/only reproductive season. In a patch-based model, the shape of patches may change, patches may
be removed and new patches may be created where there was previously inter-patch matrix. Thus some populations may be extirpated (in a non-structured population, all individuals die; in a stage-structured population,
all individuals either die or have an immediate opportunity to disperse), and new populations may arise from colonisation of newly suitable areas.

However, there are certain restrictions. Any part of the original landscape which was a ‘no-data’ region (e.g. the sea or land beyond a study area boundary) must remain in that state for the whole simulation.
The identity of patches is not cross-checked between changes, and care must therefore be taken to ensure consistency; otherwise, a patch (and its resident population) can jump to a distant location or be split into
two or more disjunct parts, with unpredictable and possibly weird consequences. It is legitimate for a patch to be split into two or more separate patches (e.g. by construction of a motorway or some other barrier),
but any existing population will remain with the part (if any) which retains the original patch number, and populations within the other parts (having a new patch number) must arise through colonisation.
Possible ways to work around this restriction include:
\itemize{
   \item Assign to all post-change parts of the original patch a new, unique patch number and specify that dispersal is allowed after population destruction (which is possible only for a structured population), in which
case some colonisation of the new patches should occur. Note that the connectivity matrix will be misleading in such cases, as every successful ‘disperser’ will appear to have moved from patch N to patch M (where M
is the new patch number).
   \item Instead of a single original patch, define two (or more) distinct but adjacent patches in the original landscape, so that they each retain their own populations when they become separated by the landscape change.
}

A dynamic landscape can be specified using the slots \code{LandscapeFile} (, \code{PatchFile}, \code{CostsFile}) and \code{DynamicLandYears}. \code{LandscapeFile} (and \code{PatchFile}, \code{CostsFile}) take a character vector with the filenames of the maps
to be loaded. All provided maps must agree in resolution, extent and origin. \code{DynamicLandYears} is a number vector that contains the years, in which these landscapes shall be loaded; it must have the same ordering so
that years and maps can be matched. If a specific map is used multiple times, it must be listed each time nevertheless.
}
\author{
Anne-Kathleen Malchow
}
