% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_InitialisationParams.R
\docType{class}
\name{Initialise}
\alias{Initialise}
\title{Set Initialisation Parameters}
\usage{
Initialise(InitType = 0, FreeType = 1, SpType = 0, NrCells, InitIndsFile = "NULL",
           InitDens = 1, IndsHaCell, PropStages = 0, InitAge = 2, minX, minY, maxX, maxY,
           InitFreezeYear = 0, RestrictRows = 0, RestrictFreq = 0, FinalFreezeYear = 0)
}
\arguments{
\item{InitType}{Type of initialisation:\cr
\code{InitType} = \eqn{0}: Free initialisation according to habitat map (default) (set \code{FreeType}), \cr
\code{InitType} = \eqn{1}: From loaded species distribution map (set \code{SpType}),\cr
\code{InitType} = \eqn{2}: From initial individuals list file (set \code{InitIndsFile}).
\cr Must to be \eqn{0} for an \code{\link[RangeShiftR]{ArtificialLandscape}}.}

\item{FreeType, NrCells}{Option for \emph{free initialisation}, i.e. required only if \code{InitType}\eqn{ = 0}:\cr
\code{FreeType} = \eqn{0}: Random; provide number of cells/patches to initialise in \code{NrCells}. \cr
\code{FreeType} = \eqn{1}: All suitable cells/patches (default).}

\item{SpType, NrCells}{Option for \emph{initialisation from species distribution map}, i.e. required only if \code{InitType}\eqn{ = 1}:\cr
\code{SpType} = \eqn{0}: All suitable cells within all distribution presence cells (default), \cr
\code{SpType} = \eqn{1}: All suitable cells within some randomly chosen presence cells; set number of cells to initialise in \code{NrCells}.}

\item{InitIndsFile}{Name of \emph{initial individuals list file}, required only if \code{InitType}\eqn{ = 2}.\cr
For informaton on the required file format see the Details below.}

\item{InitDens, IndsHaCell}{Number of individuals to be seeded in each cell/patch:\cr
\code{InitDens} = \eqn{0}: At \code{K_or_DensDep},\cr
\code{InitDens} = \eqn{1}: At half \code{K_or_DensDep} (default),\cr
\code{InitDens} = \eqn{2}: Set the number of individuals per cell/hectare to initialise in \code{IndsHaCell}.}

\item{PropStages}{For \code{\link[RangeShiftR]{StageStructure}}d models only: Proportion of individuals initialised in each stage.
Requires a vector of length equal to the number of stages; its entries must be \eqn{\ge 0} and sum to \eqn{1.0}. However, juveniles
can't be initialised and thus stage \eqn{0} (first entry) must have a value of \eqn{0.0}.}

\item{InitAge}{For \code{StageStructure}d models only: Initial age distribution within each stage:\cr
\code{InitAge} = \eqn{0}: Minimum age for the respective stage.\cr
\code{InitAge} = \eqn{1}: Age randomly sampled between the minimum and the maximum age for the respective stage.\cr
\code{InitAge} = \eqn{2}: According to a quasi-equilibrium distribution (default).}

\item{minX, maxX, minY, maxY}{Option for \emph{free initialisation} (\code{InitType}\eqn{ = 0}): Restrict initial range in X- and/or Y- coordinates, given in number of cells.
All must be \eqn{\ge 0} and (\code{maxX,maxY})\eqn{>}(\code{minX,minY}). (Integer)}

\item{InitFreezeYear}{Option for \emph{free initialisation} (\code{InitType}\eqn{ = 0}): The year until which species is confined to
its initial range limits.}

\item{RestrictRows}{Option for \emph{free initialisation} (\code{InitType}\eqn{ = 0}): The number of rows at northern front to restrict range.
If set to \eqn{0}, the range restriction turned off.}

\item{RestrictFreq}{Option for \emph{free initialisation} (\code{InitType}\eqn{ = 0}): Frequency in years at which range is restricted to northern front.}

\item{FinalFreezeYear}{Option for \emph{free initialisation} (\code{InitType}\eqn{ = 0}): The year after which species is confined to its new, current range limits, after a
period of range expansion. Will be ignored if set to \eqn{0}, otherwise must be \eqn{>} \code{InitFreezeYear}.}
}
\value{
a parameter object of class "InitialisationParams"
}
\description{
Initialisation rules define the way in which initial individuals are placed in the landscape at the start of the simulation.
There are three initialisation types:\cr
Free initialisation according to the habitat map (\code{InitType}\eqn{= 0}),
initialisation from a loaded species distribution map (\code{InitType} \eqn{= 1}),
and the initialisation from an initial individuals list (\code{InitType} \eqn{= 2}),
with additional options for each type, see the Details.\cr
Additionally, initial density and, if applicable, initial stage and age distributions can be set.
}
\details{
\emph{Initialisation Types}\cr
\itemize{
    \item \emph{Free Initialisation.} (\code{InitType}\eqn{ = 0})\cr The population is initialised according to suitable habitat in the landscape.
Either all (\code{FreeType}\eqn{=1}), or a specified number (\code{NrCells}) of randomly selected (\code{FreeType}\eqn{=0}) cells/patches
will be seeded. When using an artificial landscape, this is the only option available.
    \item \emph{From loaded species distribution map.} (\code{InitType}\eqn{ = 1})\cr The population is initialised according to a loaded
species distribution map, which needs to be provided through the module \code{\link[RangeShiftR]{ImportedLandscape}}, option \code{SpDistFile}.
All habitat cells/patches within either all (\code{SpType}\eqn{=0}), or a specified number (\code{NrCells}) of randomly selected
(\code{SpType}\eqn{=1}) presence cells (which can have a lower resolution) specified by this distribution map will seeded.
    \item \emph{From initial individuals list file.} (\code{InitType}\eqn{ = 2})\cr The population is initialised according to a list of specific
individuals (of given sex, age and stage, if appropriate) in specified cells/patches. This option allows simulation of a reintroduction
scenario.\cr The list has to be loaded from a file in the path given by \code{InitIndsFile}. It must be a tab-separated list with
explicit column headers and one row for each individual to be initialized. The expected column headers depend on the model settings and
must match the following order exactly: 'Year', 'Species' (must \eqn{= 0}), for cell-/patch-based: 'X', 'Y' / 'PatchID', 'Ninds', for sexual model: 'Sex',
for stage-structured population: 'Age', 'Stage'. The sex is specified with \eqn{0} for \emph{female} and \eqn{1} for \emph{male}.
}

\emph{Initialial density, stage, and age}\cr
For \code{InitType}\eqn{ = {0,1}}, the number of individuals that should be seeded in each cell or patch has to be set.
There are three options:
\itemize{
    \item \emph{At} \code{K_or_DensDep}. (\code{InitDens}\eqn{=0})\cr The cell/patch will be saturated at its respective \eqn{K} or \eqn{1/b}.
    \item \emph{At half} \code{K_or_DensDep}. (\code{InitDens}\eqn{=1})\cr The cell/patch will be initialised at half its \eqn{K} or \eqn{1/b}.
    \item \emph{Set value} \code{IndsHaCell}. (\code{InitDens}\eqn{=2})\cr Set the number of individuals to be seeded in each cell or the density
in each patch (in units of individuals per hectare).
}
(These settings have no effect for \code{InitType}\eqn{ = 2}.)

In the case of \code{\link[RangeShiftR]{StageStructure}}d models, the initial stage and age distributions must be specified.
If \code{InitType}\eqn{ = 2}, this is done via the \code{InitIndsFile}, whereas for \code{InitType}\eqn{ = {0,1}},
the proportion of individuals that should be initialised at each stage class is set via the numeric vector \code{PropStages}. It needs
to have as many entries as number of stages, starting from the juvenile stage (\eqn{0}). Note that these proportions must sum up to \eqn{1.0},
however the proportion of juveniles must be \eqn{0.0}.

Options for initial age distributions:
\itemize{
    \item \emph{Minimal age} of respective stage. (\code{InitAge}=0)
    \item \emph{Randomise.} (\code{InitAge}=1)\cr Individuals initialised in each stage will get an age randomly sampled between
the minimum and the maximum age for their respective stage.
    \item \emph{Quasi-equilibrium.} (\code{InitAge}=2)\cr Initial age distribution is set approximately in accordance with the number of years
taken to pass through the stage and the (female) survival rate of the respective stage.
}
(These settings have no effect for \code{InitType}\eqn{ = 2}.)

\emph{Additional options for free initialisation}\cr
In the case of free initialisation, either all or a specified number of randomly selected cells/patches will be seeded.
It is possible to restrict the landscape area available for initialisation by setting limits to the x- and y- coordinates using
\code{minX,maxX,minY} and \code{maxY}).

Additionally, for all free initialisation methods, it is possible to restrict the available landscape to the initialised range for
a specified period, specifically until year \code{InitFreezeYear}. This option may be particularly useful in an evolutionary model,
to allow a population to reach equilibrium in its original range before being allowed to expand into the remainder of the landscape.

\emph{Restrict range to northern front} is a further option provided as an additional feature for models of range expansion. Whilst not
strictly an initialisation option, it is provided in this context as it will typically be applied together with restriction of the
initial range in theoretical eco-evolutional models, where the population is firstly allowed to evolve certain traits in a stable range,
then allowed to expand its range during a period of environmental change (when dispersal traits can come under strong selection), and
finally further restricted to a stable range (set \code{FinalFreezeYear}). During the period of range expansion, the population can build up to very large numbers
of individuals in the core of the range, but the area of interest, where strong selection occurs, lies close to the range front.
Therefore, individuals in the core can be discarded, and this is achieved by progressively restricting the suitable range to a number
of rows behind the current range front (set \code{RestrictRows}) so that local populations in the core go extinct. This occurs at
a frequency \code{RestrictFreq} specified in years. The parameter \code{FinalFreezeYear} sets the year after which the range will be
frozen to its new, current limits. It must either be zero, in which case range expansion continues until either the end of the simulation or the northern edge of the
landscape is reached, or set to a year after the year specified in \code{InitFreezeYear}.

This feature is provided only for expansion in a northerly direction (increasing Y); the initial population should therefore be
established in the southern part of the landscape (low Y values). To turn it off, set \code{RestrictRows = 0}.
}
\examples{
init <- Initialise() # use all defaults (a stage-structured model requires setting PropStages)
}
\author{
Anne-Kathleen Malchow
}
