% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeShiftR.R
\docType{package}
\name{RangeShiftR-package}
\alias{RangeShiftR}
\alias{RangeShiftR-package}
\title{RangeShiftR: An R package for individual-based simulation of spatial eco-evolutionary dynamics and species' responses to environmental changes}
\description{
An R interface to the RangeShifter 2.0 simulation software.
}
\details{
\emph{RangeShiftR} offers an interface to the \emph{RangeShifter} simulation software
\insertCite{bocedi2014,bocedi2020}{RangeShiftR},
making it available for all machines that can build and run \emph{R}.

The \emph{RangeShifter} project is presented on our website
\url{https://rangeshifter.github.io/}.

To get started with the package, please find an overview and a range of tutorials here:
\url{https://rangeshifter.github.io/RangeshiftR-tutorials/}.
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rangeshifter.github.io/RangeshiftR-tutorials/}
}

}
\author{
\strong{Maintainer}: Anne-Kathleen Malchow \email{rangeshiftr@uni-potsdam.de}

Authors:
\itemize{
  \item Greta Bocedi
  \item Stephen C.F. Palmer
  \item Justin M.J. Travis
  \item Damaris Zurell
}

}
\keyword{internal}
