% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_handling.R
\name{readPop}
\alias{readPop}
\title{Read 'pop' file}
\usage{
readPop(s, dirpath, ...)
}
\arguments{
\item{s}{RSmaster parameter object}

\item{dirpath}{RS directory path}

\item{center}{In a cell-based model, the x- and y- coordinates will be converted from *nrow* and *ncol* (as in the 'pop' output file)
to true coordinates of the center of the cell measured in metres.}
}
\value{
a data.frame
}
\description{
Read the RangeShiftR output file 'pop' into a data.frame, if it was generated.
}
