% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_handling.R
\name{readRange}
\alias{readRange}
\title{Read 'range' file}
\usage{
readRange(s, dirpath)
}
\arguments{
\item{s}{RSmaster parameter object}

\item{dirpath}{RS directory path}
}
\value{
a data.frame
}
\description{
Read the RangeShiftR output file 'range' into a data.frame, if it was generated.
}
