% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DispersalParams.R
\docType{class}
\name{Settlement}
\alias{Settlement}
\title{Set Settlement Parameters}
\usage{
Settlement(StageDep = FALSE, SexDep = FALSE,
          Settle = 0, FindMate = FALSE,
          DensDep = FALSE,
          IndVar = FALSE, TraitScaleFactor,
          MinSteps = 0, MaxSteps = 0, MaxStepsYear = 0)
}
\arguments{
\item{StageDep}{Stage-dependent settlement requirements? (default: \code{FALSE})}

\item{SexDep}{Sex-dependent settlement requirements? (default: \code{FALSE})}

\item{Settle}{Settlement codes (for \code{DispersalKernel}) or settlement probability parameters (for \emph{Movement process}) for all
stages/sexes, defaults to \eqn{0} (i.e. 'die when unsuitable' for \emph{DispersalKernel} and
'always settle when suitable' for \emph{Movement process}). See the Details below.}

\item{FindMate}{Mating requirements to settle? Set for all stages/sexes. Must be \code{FALSE} (default) in a female-only model.}

\item{DensDep}{Movement process only: Density-dependent settlement probability? (default: \code{FALSE})}

\item{IndVar}{Movement process only: Individual variability in settlement probability traits? Must be \code{FALSE} (default),
if \code{DensDep=FALSE} or \code{StageDep=TRUE}.}

\item{TraitScaleFactor}{Movement process only, required if \code{IndVar=TRUE}: Scaling factors for the three traits of density-dependent settlement,
a numeric of length \eqn{3}.}

\item{MinSteps}{Movement process only: Minimum number of steps. Defaults to \eqn{0}.}

\item{MaxSteps}{Movement process only: Maximum number of steps. Must be \eqn{0} or more; set to \eqn{0} (default) for “per-step mortality only”.}

\item{MaxStepsYear}{Movement process and stage-structured population only: Maximum number of steps per year, if there are more than \eqn{1} reproductive seasons (option \code{RepSeasons} in \code{\link[RangeShiftR]{StageStructure}}).
Must be \eqn{0} or more. If \eqn{0}, every individual completes the dispersal phase in one year, i.e. between two successive reproduction phases.}
}
\value{
a parameter object of class "SettlementParams"
}
\description{
Settlement, or immigration, is the last phase of dispersal, when the organism stops in a new cell or patch of breeding habitat. The
available settlement conditions vary depending on the used \code{\link[RangeShiftR]{Transfer}} type. In any case, dispersing individuals
are not allowed to settle in their natal cell or patch and can only settle in suitable habitat.\cr
\emph{RangeShiftR} incorporates some basic settlement rules that can be stage- or sex-specific or both (set \code{StageDep}, \code{SexDep}).
If a movement process is used, density-dependence (\code{DensDep}) and/or inter-individual variability (\code{IndVar}) are available.
}
\details{
The settlement phase is determined by a suite of strategies, behaviours and reaction norms that lead individuals to the decision to stop
in a particular place. Habitat selection, mate finding and density dependence are probably three of the main processes involved,
but not the only ones. Like emigration, settlement is a complex process affected by multiple criteria including inter-individual
variability and context dependencies.

The type of implemented settlement rules depends on the movement model utilized for the \code{\link[RangeShiftR]{Transfer}}.
In any case, dispersing individuals are not allowed to settle in their natal cell or patch.\cr
\emph{RangeShiftR} incorporates some basic settlement rules that can be stage- or sex-specific or both (set \code{StageDep}, \code{SexDep}).
Inter-individual variability (\code{IndVar}) is implemented only for movement processes and then for the three traits
determining density-dependent settlement (\ifelse{html}{\out{S<sub>0</sub>}}{\eqn{S_0}}, \ifelse{html}{\out{&alpha;<sub>S</sub>}}{\eqn{α_S}},
\ifelse{html}{\out{&beta;<sub>S</sub>}}{\eqn{β_S}}; see below). In this case, settlement may not be stage-dependent.\cr

\emph{Settlement with dispersal kernels}\cr
When using a \code{\link[RangeShiftR]{DispersalKernel}}, individuals are displaced directly from the starting location to the arrival location. The suitability
of the arrival cell or patch determines whether the disperser is successful or not.\cr For species with \emph{non-overlapping generations},
where individuals have only one chance to disperse and reproduce, the model has two options if the arrival cell is unsuitable: the
individual either dies (\code{Settle} code \eqn{0}) or it can move to one of the eight neighbouring cells in the case that at least one
of them is suitable (\code{Settle} code \eqn{2}). In
the latter case, if more than one of the neighbouring cells is suitable, the individual is placed in one of them chosen randomly.
For patch-based models, if the arrival patch is unsuitable, the individual either dies or can move to a randomly chosen neighbouring
suitable patch, provided that the new patch is only one cell apart from the arrival patch.\cr For species with \emph{overlapping generations}
(i.e. a \code{\link[RangeShiftR]{StageStructure}}d population),
where individuals can disperse over multiple seasons, there are two additional options: First, if the arrival cell/patch is unsuitable,
the individual can stay there waiting until the next dispersal event when it will disperse again according to the set kernel (\code{Settle} code \eqn{1}). Second,
if both the arrival cell/patch and all eight neighbouring cells, or all eventual neighbouring patches, are unsuitable the individual
can wait in the arrival cell/patch before moving again at the next dispersal event (\code{Settle} code \eqn{3}). The arrival cell/patch is considered suitable if
it contains breeding habitat.

If the settlement condition is the same for the entire population (\code{StageDep=FALSE} and \code{SexDep=FALSE}), then the parameter
\code{Settle} takes a single integer, specifying the corresponding settlement condition code (\eqn{0,1,2} or \eqn{3}).
In case of stage- and/or sex-specific settlement conditions, \code{Settle} must be an integer matrix with the first one/two columns
stating 'stage' and/or 'sex' (\eqn{0} for \emph{female} and \eqn{1} for \emph{male}), and the last column giving the respective settlement condition.
Exactly one row for each stage/sex-combination is required, in no particular order.

Settlement condition codes: If the individuals current step ends in unsuitable habitat, it:\cr
\eqn{0} = die (default),\cr
\eqn{1} = wait (stage-structured models only),\cr
\eqn{2} = randomly choose a suitable neighbouring cell or die,\cr
\eqn{3} = randomly choose a suitable neighbouring cell or wait (stage-structured models only).\cr

Simple example for sex-dependence only: Females choose a neighbouring cell or wait, males wait: \tabular{cc}{\out{&emsp;} 0 \tab \out{&emsp;} 3 \cr \out{&emsp;} 1 \tab \out{&emsp;} 0 }

\emph{Settlement with movement processes}\cr
If individuals are dispersing by one of the two movement processes implemented (\code{\link[RangeShiftR]{SMS}} or
\code{\link[RangeShiftR]{CorrRW}}), at each step (made simultaneously) they each evaluate their current cell or patch for the
possibility of settling. This allows for the implementation of more complex settlement rules. The simplest one is that the individual
decides to stop if there is suitable habitat; this is in any case a necessary condition (set \code{DensDep=FALSE}).\cr
If a Settlement module with a constant \eqn{S_0=0} (the default) is used in a model with \emph{movement process},
it gets converted to \eqn{S_0=1.0}, i.e. 'always settle when habitat is suitable'. \cr
\cr
Furthermore, the settlement decision can be density-dependent (set \code{DensDep=TRUE}). In this case, the individual has a probability \ifelse{html}{\out{p<sub>S</sub>}}{\eqn{p_S}}
of settling in the cell or patch \eqn{i}, given by:

\ifelse{html}{\out{&emsp;&emsp; p<sub>S</sub>(i,t) = S<sub>0</sub> / ( 1 + e<sup>-&alpha;<sub>S</sub> (N(i,t) / K(i,t) - &beta;<sub>S</sub>) </sup> ) } }{\deqn{ p_S(i,t) = S_0 / ( 1 + exp[-α_S (N(i,t)/K(i,t) - β_S) ] ) } }

In the case of stage-structured models the above equation is modified to:

\ifelse{html}{\out{&emsp;&emsp; p<sub>S</sub>(i,t) = S<sub>0</sub> / ( 1 + e<sup>-&alpha;<sub>S</sub> (b(i,t) * N(i,t) - &beta;<sub>S</sub>) </sup> ) } }{\deqn{ p_S(i,t) = S_0 / ( 1 + exp[-α_S (b(i,t) N(i,t) - β_S) ] ) } }

In the first case, \eqn{K(i,t)} is the carrying capacity of the cell/patch \eqn{i} at time \eqn{t} given by \code{K_or_DensDep}.
In the latter case, \eqn{b(i,t)} represents the strength of density dependence that is given by the inverse of \code{K_or_DensDep}.\cr
Further, \ifelse{html}{\out{S<sub>0</sub>}}{\eqn{S_0}} is the maximum settlement probability,
\eqn{N(i,t)} is the number of individuals in the cell/patch \eqn{i} at time \eqn{t},
\ifelse{html}{\out{&beta;<sub>S</sub>}}{\eqn{β_S}} is the inflection point of the function and
\ifelse{html}{\out{&alpha;<sub>S</sub>}}{\eqn{α_S}} is the slope at the inflection point.\cr

Inter-individual variability \code{IndVar=TRUE} and thus evolution is implemented only for the three traits determining density-dependent settlement
(\code{DensDep=TRUE}), and if so, it may not be stage-dependent (\code{StageDep=FALSE}).
For each trait the initial distribution in the population (as mean and standard variation) must be set in \code{Settle} (instead of only one constant value),
as well as their scaling factors in \code{TraitScaleFactor} (see \code{\link[RangeShiftR]{Genetics}}).

The parameters that determine the settlement probabilities have to be provided via the parameter \code{Settle}, which generally takes a numeric matrix, or - if only a single constant probability is
used (i.e. \code{DensDep, IndVar, StageDep, SexDep = FALSE}) - a single numeric.
The format of the matrix is defined as follows: The number of columns depend on the options \code{DensDep} and \code{IndVar}. If \code{DensDep=FALSE}, the
density-independent probability \ifelse{html}{\out{p<sub>S</sub>}}{\eqn{p_S}} must be specified. If \code{DensDep=TRUE}, the functional parameters \ifelse{html}{\out{S<sub>0</sub>}}{\eqn{S_0}},
\ifelse{html}{\out{&alpha;<sub>S</sub>}}{\eqn{α_S}} and \ifelse{html}{\out{&beta;<sub>S</sub>}}{\eqn{β_S}} (cf. equation above) must be specified.
Additionally, if \code{IndVar=FALSE}, these traits are fixed, but if \code{IndVar=TRUE} each of them is replaced by two parameters: their respective initial mean and
standard deviation. They are used to normally distribute the traits values among the individuals of the initial population. Additionally, the \code{TraitScaleFactor} of
these traits have to be set.

All parameters have to be given for each stage/sex if the respective dependency is enabled. If \code{StageDep=TRUE}, state the corresponding stage in the first column.
If \code{SexDep=TRUE}, state the corresponding stage in the next (i.e. first/second) column, with \eqn{0} for \emph{female} and \eqn{1} for \emph{male}.
The following table lists the required columns and their correct order for different settings:

\tabular{ccccc}{DensDep \tab IndVar \tab StageDep \tab SexDep \tab columns \cr
 F \tab F \tab F \tab F \tab - not applicable - \cr
 F \tab F \tab T \tab F \tab stage \cr
 F \tab F \tab F \tab T \tab sex \cr
 F \tab F \tab T \tab T \tab stage, sex \cr
 T \tab F \tab F \tab F \tab \ifelse{html}{\out{S<sub>0</sub>}}{\eqn{S_0}}, \ifelse{html}{\out{&alpha;<sub>S</sub>}}{\eqn{α_S}}, \ifelse{html}{\out{&beta;<sub>S</sub>}}{\eqn{β_S}} \cr
 \out{&#8942;} \tab \out{&#8942;} \tab \out{&#8942;} \tab \out{&#8942;} \tab \out{&#8942;} \cr
 T \tab F \tab T \tab T \tab stage, sex, \ifelse{html}{\out{S<sub>0</sub>}}{\eqn{S_0}}, \ifelse{html}{\out{&alpha;<sub>S</sub>}}{\eqn{α_S}}, \ifelse{html}{\out{&beta;<sub>S</sub>}}{\eqn{β_S}} \cr
 T \tab T \tab F \tab F \tab mean\ifelse{html}{\out{(S<sub>0</sub>)}}{\eqn{(S_0)}}, sd\ifelse{html}{\out{(S<sub>0</sub>)}}{\eqn{(S_0)}}, mean\ifelse{html}{\out{(&alpha;<sub>S</sub>)}}{(\eqn{α_S})}, sd\ifelse{html}{\out{(&alpha;<sub>S</sub>)}}{(\eqn{α_S})}, mean\ifelse{html}{\out{(&beta;<sub>S</sub>)}}{(\eqn{β_S})}, sd\ifelse{html}{\out{(&beta;<sub>S</sub>)}}{(\eqn{β_S})} \cr
 T \tab T \tab F \tab T \tab sex, mean\ifelse{html}{\out{(S<sub>0</sub>)}}{\eqn{(S_0)}}, sd\ifelse{html}{\out{(S<sub>0</sub>)}}{\eqn{(S_0)}}, mean\ifelse{html}{\out{(&alpha;<sub>S</sub>)}}{(\eqn{α_S})}, sd\ifelse{html}{\out{(&alpha;<sub>S</sub>)}}{(\eqn{α_S})}, mean\ifelse{html}{\out{(&beta;<sub>S</sub>)}}{(\eqn{β_S})}, sd\ifelse{html}{\out{(&beta;<sub>S</sub>)}}{(\eqn{β_S})}
 }

The column headings need not be included, only the numeric matrix is required. The rows require no particular order, but there must be exactly one row for each stage/sex combination.
For example, in the case of density-, stage- and sex-dependent settlement with no individual variability:
\tabular{ccccc}{ \out{&emsp;} 0 \tab \out{&emsp;} 0 \tab \out{&emsp;} 1.0 \tab \out{&emsp;} 0.2 \tab \out{&emsp;} 4.0 \cr
 \out{&emsp;} 0 \tab \out{&emsp;} 1 \tab \out{&emsp;} 1.0 \tab \out{&emsp;} 0.1 \tab \out{&emsp;} 6.0 \cr
 \out{&emsp;} 1 \tab \out{&emsp;} 0 \tab \out{&emsp;} 0.7 \tab \out{&emsp;} 0.5 \tab \out{&emsp;} 2.0 \cr
 \out{&emsp;} 1 \tab \out{&emsp;} 1 \tab \out{&emsp;} 0.5 \tab \out{&emsp;} 0.5 \tab \out{&emsp;} 2.0 \cr
 \out{&emsp;} 2 \tab \out{&emsp;} 0 \tab \out{&emsp;} 0.05 \tab \out{&emsp;} 1.0 \tab \out{&emsp;} 1.0 \cr
 \out{&emsp;} 2 \tab \out{&emsp;} 1 \tab \out{&emsp;} 0.05 \tab \out{&emsp;} 1.0 \tab \out{&emsp;} 1.0
}

To avoid having individuals moving perpetually because they cannot find suitable conditions to settle, the model requires a maximum number
of steps (\code{MaxSteps}) or a per-step mortality (within the \code{Transfer} object), or both, to be set.
The maximum number of steps defines the maximum time length of the transfer period.
When an individual reaches the maximum number of steps, it stops where it is regardless of the suitability of the location. In
the case of \emph{non-overlapping generations} this results in automatic death if the individual stops in unsuitable habitat. For species that can disperse
over multiple seasons, i.e. a \emph{stage-structured population}, the model additionally requires a maximum number of steps per dispersal event (\code{MaxStepsYear}); on reaching that limit, the individual will
stop where it is and the next season, if still alive, it will move again.

An additional rule that can be set constitutes a minimum number of steps (\code{MinSteps}) that each individual must take before settlement can take
place. This is useful for simulating situations where animals, in a ‘dispersal mode’, will keep moving and not consider settling even
if suitable conditions are available \insertCite{@e.g. @barton2012risky}{RangeShiftR}.

\emph{Mating requirements}\cr
Sexual species may be required to find a mate, i.e. there has to be at least one individual of the opposite sex present for the cell/patch to be considered suitable for settlement.
Density-dependence and mating requirements can also be combined together to determine the settlement decision.

The application of this mating condition can be switched on or off using the parameter \code{FindMate}, which takes either a single
logical if \code{StageDep=FALSE} and \code{SexDep=FALSE} or, otherwise, a logical vector of same length as the number rows in
\code{Settle}, using same order.
}
\references{
\insertAllCited{}
}
\author{
Anne-Kathleen Malchow
}
