% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SimulationParams.R
\docType{class}
\name{Simulation}
\alias{Simulation}
\title{Set Simulation parameters}
\usage{
Simulation(Simulation = 1, Replicates = 2, Years = 50, Absorbing = FALSE,
           Gradient = 0, GradSteep, Optimum, f, ExtinctOptim,
           Shifting = FALSE, ShiftRate, ShiftStart, ShiftEnd,
           LocalExt = FALSE, LocalExtProb,
           EnvStoch = 0, EnvStochType, std, ac, minR, maxR, minK, maxK,
           OutIntRange = 1, OutIntOcc = 0,
           OutIntPop = 1, OutIntInd = 0,
           OutIntTraitCell = 0, OutIntTraitRow = 0,
           OutIntConn = 0, OutIntPaths = 0, OutIntGenetic = 0,
           OutGenType = 0, OutGenCrossTab = FALSE,
           OutStartPop = 0, OutStartInd = 0,
           OutStartTraitCell = 0, OutStartTraitRow = 0,
           OutStartConn = 0, OutStartPaths = 0, OutStartGenetic = 0,
           SMSHeatMap = FALSE)
}
\arguments{
\item{Simulation}{ID number of current simulation, defaults to \eqn{1}. (integer)}

\item{Replicates}{Number of simulation iterations, defaults to \eqn{2}. (integer)}

\item{Years}{The number of simulated years, defaults to \eqn{50}. (integer)}

\item{Absorbing}{If \code{FALSE} (default), every move in the \code{\link[RangeShiftR]{Transfer}} process will be
repeated until a valid cell is met.\cr
If \code{TRUE}, an individual which hits a non-valid cell or
transgresses the landscape boundary during the dispersal act is eliminated from the simulation.}

\item{Gradient}{Whether to apply north-south gradient:
\eqn{0} = None (default), \cr
\eqn{1} = decreasing \code{K_or_DensDep}\cr
\eqn{2} = decreasing growth rate \eqn{r} or, for a \code{\link[RangeShiftR]{StageStructure}}d population,
fecundity \ifelse{html}{\out{&phi;}}{\eqn{\phi}}, \cr
\eqn{3} = increasing local extinction probability \eqn{e}. \cr
If activated, a gradient will be imposed along the north-south (\eqn{y}-) axis in which the selected parameter varies linearly with distance from the
optimum \eqn{y}-value \code{Optimum}. Note that a \code{Gradient} can not be applied in for patch-based models (must be \code{Gradient}\eqn{=0}).}

\item{GradSteep}{Required if \code{Gradient} \ifelse{html}{\out{&ne; 0}}{\eqn{> 0}}: gradient steepness in units of (fraction of local value) per cell. Must be \eqn{\ge 0}.}

\item{Optimum}{Required if \code{Gradient} \ifelse{html}{\out{&ne; 0}}{\eqn{> 0}}: \eqn{y}-value at which the extremum is obtained. Must be \eqn{\ge 0}.}

\item{f}{Required if \code{Gradient} \ifelse{html}{\out{&ne; 0}}{\eqn{> 0}}: local scaling factor that determines the magnitude of stochastic local heterogeneity relative to the optimal value. Must be \eqn{\ge 0}.}

\item{ExtinctOptim}{Required if \code{Gradient} \eqn{= 3}: optimum (i.e. minimal) local extinction probability at \code{Optimum}. Must be between \eqn{0} and \eqn{1}.}

\item{Shifting}{Only applicable if \code{Gradient} \ifelse{html}{\out{&ne; 0}}{\eqn{> 0}}:\cr
If \code{FALSE} (default), the gradient is stationary.\cr
If \code{TRUE}, the \code{Gradient} shifts along the \eqn{y}-axis towards increasing \eqn{y} (northwards). Requires to set \code{ShiftRate}, \code{ShiftStart} and \code{ShiftEnd},}

\item{ShiftRate}{Required if \code{Shifting=TRUE}: shift rate of the gradient in units of rows per year. (integer)}

\item{ShiftStart}{Required if \code{Shifting=TRUE}: year in which the gradient shifting starts (integer)}

\item{ShiftEnd}{Required if \code{Shifting=TRUE}: year in which the gradient shifting stops. (integer)}

\item{LocalExt}{If \code{FALSE} (default), no additional extinction probability is applied.\cr
If \code{TRUE}, an independent constant extinction probability \code{LocalExtProb} is applied, defined as
the probability that each population goes extinct at each year.
Note that \code{LocalExt} must be \code{FALSE} for a patch-based model or if \code{Gradient}\eqn{=3}.}

\item{LocalExtProb}{Required if \code{LocalExt=TRUE}: independent yearly extinction probability of populations.}

\item{EnvStoch}{Scale of environmental stochasticity:\cr
\eqn{0} = none (default),\cr
\eqn{1} = global (a single time series for entire landscape),\cr
\eqn{2} = local (each cell fluctuates independently, only permitted for cell-based model).\cr
Environmental stochasticity is always applied on a yearly basis.}

\item{EnvStochType}{Required if \code{EnvStoch} \ifelse{html}{\out{&ne; 0}}{\eqn{> 0}}: Parameter to which environmental stochasticity is applied:\cr
\eqn{0} = growth rate \eqn{r} or, for a \code{\link[RangeShiftR]{StageStructure}}d population, fecundity (\eqn{\phi}).\cr
\eqn{1} = demographic density dependence \code{K_or_DensDep} (carrying capacity or 1/b) (allowed for artificial landscapes only!).}

\item{std}{Required if \code{EnvStoch} \ifelse{html}{\out{&ne; 0}}{\eqn{> 0}}: magnitude of stochastic fluctuations. Must be \eqn{> 0.0} and \eqn{\le 1.0}.}

\item{ac}{Required if \code{EnvStoch} \ifelse{html}{\out{&ne; 0}}{\eqn{> 0}}: temporal autocorrelation coefficient. Must be \eqn{\ge 0.0} and \eqn{<1.0}.}

\item{minR, maxR}{Required if \code{EnvStochType}\eqn{=0}: minimum and maximum growth rates.}

\item{minK, maxK}{Required if \code{EnvStochType}\eqn{=1}: minimum and maximum value of \eqn{K} or \eqn{1/b}, respectively.}

\item{OutIntRange, OutIntOcc, OutIntPop, OutIntInd, OutIntGenetic, OutIntTraitCell, OutIntTraitRow, OutIntConn, OutIntPaths}{Control the various types
of Output files, i.e. range, occupancy, populations, individuals, traits (by cell or by row), connectivity, SMS paths and genetics:\cr
 \eqn{=0 }: Output disabled.\cr
 \eqn{>0 }: Output enabled; sets interval (in years) in which output is generated.\cr
If the output is enabled, start values are required. By default, only the output of Range and Population are enabled.\cr
Occupancy output is only applicable if \code{Replicates>1}.
Traits output is only applicable for a cell-based model with inter-individual variability.
Connectivity output is only applicable for a patch-based model.
SMS paths is only applicable for a model with SMS transfer method.}

\item{OutStartPop, OutStartInd, OutStartGenetic, OutStartTraitCell, OutStartTraitRow, OutStartConn, OutStartPaths}{Starting years for output generation. Note that the first year is year \eqn{0}. Defaults to \eqn{0} for all output types. (integer)}

\item{OutGenType}{Required if \code{OutIntGenetic}\eqn{>0}: Genetics output will be generated for:\cr
\eqn{0} = juveniles only (default), \eqn{1} =  all individuals, \eqn{2} = adults only.}

\item{OutGenCrossTab}{Required if \code{OutIntGenetic}\eqn{>0}:\cr
If \code{FALSE} (default), Genetics output will be written to several files.
\cr If \code{TRUE} Genetics output is generated as a cross table.}

\item{SMSHeatMap}{Produce SMS heat map raster as output? Defaults to \code{FALSE}.}
}
\description{
Set basic simulation parameters and control output types.\cr
Furthermore, optionally define a (\code{Shifting}) Environmental \code{Gradient}, Environmental Stochasticity (\code{EnvStoch}) and/or Local extinction (\code{LocalExt}).
(These options are to be moved to separate classes in future versions.)
}
\details{
\emph{Environmental Gradient}\cr
In \emph{RangeShiftR}, it is possible to superimpose an artificial gradient on top of the landscape map (real or artificial).
Gradients are implemented for cell-based models only.\cr
An environmental gradient can be superimposed on the habitat map to describe gradual change in abiotic factors through space. Use the option \code{Gradient}
and choose one of three implemented parameter gradients. These are, respectively for non-structured / stage-structured population models:\cr
\itemize{
   \item{Decreasing values of \code{K_or_DensDep}, that mediates demographic density dependence (stronger with lower values) and is interpreted as
         the \emph{carrying capacity} \eqn{K} / the \emph{strength of density dependence} \ifelse{html}{\out{b<sup>-1</sup>}}{\eqn{1/b}} (set \code{Gradient=1});}
   \item{Decreasing growth rate \eqn{r} / fecundity (\eqn{\phi}) (set \code{Gradient=2});}
   \item {Increasing local extinction probability \eqn{e} (set \code{Gradient=3}).}
}
The gradient is restrictively implemented along the north-south (\eqn{y})-axis and the selected parameter declines linearly with (\eqn{y})-distance from
an optimum location (\code{Optimum}).

Gradients are implemented following the method of \insertCite{travis2004;textual}{RangeShiftR}, which combines linear variability with local heterogeneity.
If \eqn{Z} is one of the gradient variables listed above, \eqn{Z={K, 1/b, r, \phi, e}}, the value of \eqn{Z(x,y)} for a cell with \eqn{x} and \eqn{y}-coordinates
is given by the following equation:

\ifelse{html}{\out{&nbsp; &nbsp; Z(x,y) = Z<sub>0</sub> * z(x,y) &nbsp; &nbsp; &nbsp; for K and r,}}{\deqn{Z(x,y)=Z_0 * z(x,y)} for K and r,}
or\cr
\ifelse{html}{\out{&nbsp; &nbsp; Z(x,y) = 1 - z(x,y) + e<sub>opt</sub> &nbsp; &nbsp; for e}}{\deqn{Z(x,y)= 1 - z(x,y) + e_{opt}} for e}
\cr

with \cr
\ifelse{html}{\out{&nbsp; &nbsp; z(x,y) = 1 - |y - y<sub>opt</sub>| G + U(-1,1) f }}{\deqn{z(x,y) = 1 - |y - y_{opt}| G + U(-1,1) f }}
, constrained to \eqn{\ge 0}

where \ifelse{html}{\out{Z<sub>0</sub>}}{Z_0} is the original parameter value at \eqn{(x,y)},
\ifelse{html}{\out{e<sub>opt</sub>}}{e_{opt}} (\code{ExtinctOptim}) the minimum extinction probability.
The linear variability is specified by
\ifelse{html}{\out{y<sub>opt</sub>}}{y_{opt}} (\code{Optimum}), the \eqn{y}-value at which the extremum (i.e. \ifelse{html}{\out{Z<sub>0</sub>}}{Z_0} or \ifelse{html}{\out{e<sub>opt</sub>}}{e_{opt}}) is obtained,
and \eqn{G} (\code{GradSteep}), the gradient steepness in units of fraction of the local value per cell.
The local heterogeneity is determined by a random number drawn from a uniform distribution between \eqn{-1} and \eqn{1} for each cell
and \code{f}, the local scaling factor that determines the magnitude of this stochastic local variation relative to the extremal value.

The gradient in fecundity φ applies to the fecundity of each stage. Negative local values in \eqn{z(x,y)} are set to \eqn{0}.

It is also possible to simulate the shifting of the gradient by setting the option \code{Shifting}. Here the position \eqn{y} of the species’
optimum is shifted northwards (increasing \eqn{y}) at a given rate \code{ShiftRate} (in units of rows per year),
starting from year \code{ShiftStart} to year \code{ShiftEnd}.

Environmental gradients are available for cell-based models only, due to the cell-based character of operations and therefore \code{Gradient}
has to be \eqn{0} for patch-based models.

\emph{Local Extinctions}\cr
An additional, independent extinction probability can be added using the option \code{LocalExt}. If set,
in each year, every population has an identical probability \code{LocalExtProb} of going extinct.
This does not affect any demographic parameters but simply kills off the local population.

\emph{Environmental Stochasticity}\cr
It is possible to model environmental stochasticity via the option \code{EnvStoch} acting at a global or local scale
and can be applied to \code{K_or_DensDep}, the demographic density dependence (\code{EnvStoch=1}), or to growth rate / fecundity (\code{EnvStoch=0}).
It is implemented using a first order autoregressive process to generate time series of the noise value \eqn{ε}
\insertCite{ruokolainen2009}{RangeShiftR}:

\deqn{ε(t+1) = κ ε(t) + \omega(t) \sqrt(1-\kappa^2)}

where κ is the autocorrelation coefficient (\code{ac}) and ω is a random normal variable drawn from \eqn{N(0,σ)}.
Changing σ (\code{std}) changes the magnitude of the fluctuations. The spatial scale of the variation can either be global (a single time series
for the entire landscape) or local (each cell fluctuates independently), and is always applied on a yearly basis.
Different degrees of spatial autocorrelation are not implemented in the current version.

The noise affects the species' selected parameter \eqn{Z} as follows:
\deqn{Z(x,y,t) = Z(x,y,0) + Z ε(t)}
where \eqn{x} and \eqn{y} are the cell coordinates and \eqn{Z} is the original parameter value in absence of stochasticity and gradients.
In the presence of an environmental gradient, \eqn{Z(x,y,0)} is the gradient value at the cell location, otherwise its equal to \eqn{Z}.
The resulting values \eqn{Z(x,y,t)} are limited to the maximum and minimum values \code{minR,maxR} or \code{minK,maxK}, respectively.

\emph{Output files}\cr
Seven different types of outputs can be produced, plus one more for patch-based models.
All the output files will be named with a standard name reporting the simulation ID number and
the type of output. The file name will start with the batch number, and also indicate the number
of the landscape to which the output refers. Additionally, for each simulation all the set parameters
will be automatically written to a text file named \"Sim0_Parameters.txt\" in the case of simulation\eqn{#=0}.

- \emph{Species range} (\code{Sim0_Range.txt}) \cr
contains the following general information regarding the species’ range:\cr
Replicate number (Rep), Year (Year), Reproductive season within the year (RepSeason), Total number of individuals (NInds),
Total number of individuals in each stage (NInd_stageX; only in case of stage-structured models),
Total number of juveniles born (NJuvs; only in case of stage-structured models),
Total number of occupied cells (NOccupCells) or total number of occupied patches (NOccupPatches),
Ratio between occupied and suitable cells or patches (OccupSuit),
Species’ range, in terms of maximum and minimum coordinates (min_X, max_X, min_Y, max_Y).\cr
Data are written before reproduction at each reproductive season at the specified yearly interval. An extra line is written
at the end of the simulation.

- \emph{Occupancy} \cr
reports the cell/patch probability of occupancy. This is only possible if the number of replicates is greater than \eqn{1}.
Two files will be produced:\cr

     1) \code{Sim0_Occupancy.txt}:  contains a list of all the cells in the landscape (\eqn{x-} and \eqn{y-}coordinates) or
of all the patches (PatchID). The remaining columns give the occupancy probability of the cell/patch at defined time steps.
The occupancy probability is obtained by dividing the number of times (replicates) that the cell/patch has been occupied in
a given year, by the total number of replicates.

     2) \code{Sim0_Occupancy_Stats.txt}: Summary occupancy statistics, i.e. the mean ratio between occupied and suitable cells
(Mean_OccupSuit) and its standard error (Std_error) at the set time interval.

    Data will be recorded at the beginning of the year before any other process (and only once a year no matter the number
of reproductive seasons per year).

- \emph{Populations} (\code{Sim0_Pop.txt}) \cr
contains statistics regarding each population present in the landscape at a given time interval:\cr
Replicate number (Rep), Year (Year), Reproductive season within the year (RepSeason), Cell location (\eqn{x-} and \eqn{y-}coordinates) or
patch ID (PatchID), Species number (Species; not yet used, always \eqn{0}), Number of individuals in the population (NInd),
Number of individuals in each stage (NInd_stageX; only in case of stage-structured models). If the reproduction is sexual,
these columns will be replaced by the number of females (Nfemales_stageX) and of males (Nmales_stageX) in each stage. In the case
of sexual model without stage structure, two columns will indicate the number of females (Nfemales) and of males (Nmales) in
the population. In the case of a stage-structured population, the number of juveniles born (NJuvs). If the reproduction is sexual,
these columns will be replaced by the number of females juveniles (NJuvFemales) and males (NJuvMales).\cr
As for the species’ range output, data are collected before reproduction at each reproductive
season at the specified yearly interval and at the end of the simulation.

- \emph{Individuals} (\code{Sim0_Rep0_Inds.txt}) \cr
contains information regarding each individual at a given time step. To avoid the production of huge files, a separate file is
saved for each replicate. Data are recorded after settlement and before aging (in the case of overlapping generations). For each
individual the following data are saved: \cr
Replicate number (Rep), Year, Reproductive season within the year (RepSeason), Species ID (always \eqn{0}),
Individual ID (IndID), the individual’s Status (Status), Natal cell (Natal_X and Natal_Y) and current cell (\eqn{x} and
\eqn{y}) coordinates or natal and current patch IDs (Natal_patch and PatchID), Sex (0 = female, 1 = male), Age in years
(in case of overlapping generations), Stage (in case of stage structure), Emigration traits, Transfer traits (depending on
transfer method).

- \emph{Genetics} (\code{Sim0_Rep0_Genetics.txt}) \cr
lists the full genome of each individual selected for output (i.e. all individuals if the population is not structured) during the reporting year
(or present in the initial population at year \eqn{0}) for the current replicate. This file can therefore be \emph{extremely large}, and should be
produced only for temporally short simulations, small populations or at infrequent reporting time intervals. It comprises:\cr
    - Replicate number (Rep), Year, Species ID (always \eqn{0}), Individual ID (IndID), \cr
and then \emph{either} one or more lines listing:
    - Chromosome number (starting from 0), Locus on this chromosome (starting from 0), value of the only allele at the locus for a
    haploid species (Allele0) or the values of both alleles at the locus for a diploid species (Allele0,Allele1) \cr
\emph{or} a single line of:
    - a set of columns having compound headings of the form \code{Chr0Loc0Allele0} derived from each chromosome, locus and allele (as above).

- \emph{Traits} \cr
In the case of inter-individual variability and evolution of the dispersal traits, it is possible to output the mean traits of
the population. There are two types of traits output:\cr

    1) \code{Sim0_TraitsXcell.txt / Sim0_TraitsXpatch.txt} reports mean and standard deviation of the varying/evolving traits for each
cell/patch, for each replicate and reproductive season at the set year interval.\cr

    2) \code{Sim0_TraitsXrow.txt} mean and standard deviation of the varying/evolving traits computed at the row (\eqn{y}) level,
pulling together all the populations occupying cells in \eqn{y}. Values are reported for each replicate and reproductive season
at the specified yearly interval. This is particularly useful for analyzing the structuring of traits along latitudinal gradients.
It is possible to compute this output only for cell-based models. \cr

    Data for these outputs are collected at the same time as for the
range and population outputs, i.e. before reproduction at each reproductive season at the set year interval and at the end of the
simulation. For sexual models, the standard deviation relates to the variation between all alleles in the local population (which
is greater than the variation in phenotypic expression; if the phenotypic s.d. is required, it must be calculated from
individual-level output data).

- \emph{Connectivity matrix} (\code{Sim0_Connect.txt}) \cr
is available for a patch-based model only. It presents counts of the number of individuals successfully dispersing from each patch
to each other patch for each year specified by \code{OutIntConn}, starting from \code{OutStartConn}. If there is more than one
reproductive season during the year, cumulative year-end totals are reported. Although the file contains the data required for
true \eqn{NxN} matrices, the data are presented in list format:\cr
Replicate number (Rep), Year (Year), ID number of natal patch (StartPatch), ID number of settlement patch (EndPatch), Number of
individuals dispersing from StartPatch to EndPatch (NInds).

- \emph{SMS paths} (\code{Sim0_Rep0_SMSpaths.txt}) \cr
is available for a model with transfer method SMS only. It lists the cell-based trajectories of all (successfully or unsuccessfully)
dispersed individuals from the natal to the final (settlement or fatal) patch for each year specified by \code{OutIntPaths}, starting
from \code{OutStartPaths}. The data are presented in list format with the columns:\cr
Year (Year), Individual ID (IndID), consecutive step number (Step), coordinates of cell at this step (\eqn{x} and \eqn{y}),
status of individual (Status).
The status is an integer number that codes for the following possible states:\cr
   0 = natal patch,\cr
   1 = disperser,\cr
   2 = disperser awaiting settlement in possible suitable patch,\cr
   3 = waiting between dispersal events,\cr
   4 = completed settlement,\cr
   5 = completed settlement in a suitable neighbouring cell,\cr
   6 = died during transfer by failing to find a suitable patch (includes exceeding maximum number of steps or crossing absorbing boundary),\cr
   7 = died during transfer by constant, step-dependent, habitat-dependent or distance-dependent mortality,\cr
   8 = failed to survive annual (demographic) mortality,\cr
   9 = exceeded maximum age.\cr\cr

- \emph{SMS Heat map} (\code{OutputMaps/Sim0_Land0_Rep0_Visits.txt}) \cr
When the transfer model is \emph{SMS}, an additional optional output is a series of maps in ASCII raster format, showing how many times each
cell has been visited by a dispersing individual across the whole time period of the simulation. These heat maps may be useful, for example,
for identifying corridors which are heavily used during the dispersal phase. One raster map is created in the \emph{Output_Maps} folder for
each replicate simulation, and is in the same format as the input habitat file.

- \emph{Log file} (\code{Batch1_RS_log.csv}) \cr
An additional log file will be created automatically. In it is listed the time taken (in seconds) to run the simulation.
It may also possibly include error codes, which can occur in rare occasions when the batch input files are in themselves valid,
but there is an inconsistency between files or an invalid habitat code or patch number occurs in an input map file.
Error codes are listed in the \emph{Batch_error_codes.xlsx} file.
}
\references{
\insertAllCited{}
}
\author{
Anne-Kathleen Malchow
}
