% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DispersalParams.R
\docType{class}
\name{SMS}
\alias{SMS}
\title{Set up a Stochastic Movement Simulator}
\usage{
SMS(PR = 1, PRMethod = 1, MemSize = 1,
    DP = 1.0,
    GoalType = 0,
    GoalBias = 1.0, AlphaDB, BetaDB,
    IndVar = FALSE,
    Costs, StepMort = 0.0,
    StraightenPath = TRUE)
}
\arguments{
\item{PR}{Perceptual range. Given in number of cells, defaults to \eqn{1}. (integer)}

\item{PRMethod}{Method to evaluate the effective cost of a particular step from the landscape within the perceptual range:\cr \eqn{1 = }Arithmetic mean (default)\cr \eqn{2 = }Harmonic
mean\cr \eqn{3 = }Weighted arithmetic mean}

\item{MemSize}{Size of memory, given as the number of previous steps over which to calculate current direction to apply directional persistence
(\code{DP}). A maximum of \eqn{14} steps is supported, default is \eqn{1}. (integer)}

\item{DP}{Directional persistence. Corresponds to the tendency to follow a correlated random walk, must be \eqn{\ge 1.0}, defaults to \eqn{1.0}.\cr
If \code{IndVar=TRUE}, expects a vector of length three specifying (Mean, SD, TraitScaleFactor) of \code{DP}.}

\item{GoalType}{Goal bias type: \eqn{0 = } None (default), \eqn{2 = } Dispersal bias.}

\item{GoalBias}{Only if \code{GoalType=2}: Goal bias strength. Must be must be \eqn{\ge 1.0}, defaults to \eqn{1.0}. \cr If \code{IndVar=TRUE}, expects a vector of length three
specifying (Mean, SD, TraitScaleFactor) of \code{GoalBias}.}

\item{AlphaDB}{Required if \code{GoalType=2}: Dispersal bias decay rate. Must be must be \eqn{> 0.0}.\cr If \code{IndVar=TRUE}, expects a vector of length three
specifying (Mean, SD, TraitScaleFactor) of \code{AlphaDB}.}

\item{BetaDB}{Required if \code{GoalType=2}: Dispersal bias decay inflection point (given in number of steps). Must be must be \eqn{> 0.0}.\cr If \code{IndVar=TRUE},
expects a vector of length three specifying (Mean, SD, TraitScaleFactor) of \code{BetaDB}.}

\item{IndVar}{Individual variability in SMS traits (i.e. \code{DP}, \code{GoalBias}, \code{AlphaDB} and \code{BetaDB})? Defaults to \code{FALSE}.}

\item{Costs}{Describes the landscapes resistance to movement. Set either: \cr
 - \emph{habitat-specific} costs for each habitat type, or\cr
 - \code{"file"}, to indictae to use the \emph{cost raster} map(s) specified in the landscape module and import the cost values from them.\cr
In the first case of \emph{habitat-specific} costs a numeric vector is expected with, respectively, length \code{Nhabitats} for an \code{\link[RangeShiftR]{ImportedLandscape}}
with habitat codes (i.e. \code{HabPercent=FALSE})) or length \eqn{2} for an \code{\link[RangeShiftR]{ArtificialLandscape}} (matrix and habitat costs).\cr
In the second case of importing a \emph{cost raster} file, specify the file name(s) in the \code{\link[RangeShiftR]{ImportedLandscape}} module.
The specified map has to match the landscape raster in extent, coordinates and resolution, and each cell contains a cost value (\eqn{\ge 1}).
This is the only option for an imported landscape with habitat qualities / cover percentage (i.e. \code{HabPercent=TRUE}).}

\item{StepMort}{Per-step mortality probability. Can be either \emph{constant}, in which case a single numeric is expected (the default, with
value \eqn{0.0}) or \emph{habitat-specific}, in which case a numeric vector (with a length as described above for \code{Costs}) is expected.
All values must be within the half-open interval \eqn{[0,1)}. \cr
For an imported habitat quality landscape (\code{HabPercent=TRUE}), only constant per-step mortality is allowed.}

\item{StraightenPath}{Straighten path after decision not to settle in a patch? Defaults to \code{TRUE}, see Details below.}
}
\value{
a parameter object of class "StochMove"
}
\description{
A method to describe \code{\link[RangeShiftR]{Transfer}}:
SMS is a stochastic individual-based movement model where organisms move through grid-based, heterogeneous landscapes. The model uses similar
cost surfaces as the least cost path (LCP) method, but it relaxes two of its main assumptions: Firstly, individuals are not assumed to be
omniscient, but move according to what they can perceive of the landscape within their perceptual range (\code{PR}). Secondly, individuals
do not know a priori their final destination, which is a reasonable assumption for dispersing individuals. For a complete description of the
method, see the Details below or refer to \insertCite{palmer2011introducing;textual}{RangeShiftR}.
}
\details{
SMS is a stochastic individual-based model where organisms move through grid-based, heterogeneous landscapes. The model uses similar cost
surfaces as the least cost path (LCP) \insertCite{adriaensen2003application,chardon2003incorporating,stevens2006gene,driezen2007evaluating}{RangeShiftR},
but it relaxes two of the main assumptions/limitations of the latter. Firstly, individuals are not assumed to be omniscient, but move according to what they
can perceive of the landscape within their perceptual range. Secondly, individuals do not know a priori their final destination, which is a
reasonable assumption for dispersing individuals. Here, the core components of SMS are briefly described;
see \insertCite{palmer2011introducing;textual}{RangeShiftR} for a complete description of the method.

SMS uses cost maps where a relative cost to movement is assigned to each habitat type. Costs are integer numbers and represent the cost of
moving through a particular land cover relative to the cost of moving through breeding habitat (which is conventionally set to a cost of \eqn{1}).
Individuals take single cell steps basing their decisions on three parameters: their perceptual range (\code{PR}) (given in number of cells),
the method used to evaluate the landscape within their perceptual range (\code{PRMethod}), and their directional persistence (\code{DP}), which corresponds to
their tendency to follow a correlated random walk. At each step, the individual evaluates the surrounding habitat in order to determine
the effective cost of taking a particular step to each of the eight neighbouring cells. The effective cost is a mean of the cost of the
neighbouring cell and the surrounding cells beyond it within the \code{PR}, and is calculated by one of three possible methods:
- \emph{Arithmetic mean}\cr
- \emph{Harmonic mean} - The reciprocal of the arithmetic mean of the reciprocals of the observations (cell costs). This method increases the
detectability of low cost cells but performs less well than the arithmetic mean in detecting high cost cells. Therefore, the choice between
the two depends on whether the main driver of the animal movement is selecting for good habitat or avoiding costly habitat.\cr
- \emph{Weighted arithmetic mean} - The cost of each cell is weighted by its inverse distance from the individual (which is assumed to be in
the centre of the current cell).

The effective cost of each neighbouring cell is weighted by the \code{DP}, which is lowest in the direction of travel. \code{DP} can be
calculated over more steps than just the previous one (up to a maximum of \eqn{14}), which is controlled by the memory size parameter (\code{MemSize})
\insertCite{palmer2014inter,aben2014simple}{RangeShiftR}. Increasing the memory size means that an individual retains for longer its tendency to move in a
certain direction, and hence paths tend to become somewhat smoother.

There is an option to include goal bias, i.e. a tendency to move towards a particular destination  \insertCite{aben2014simple}{RangeShiftR},
which is implemented in a similar way to \code{DP}. However, as dispersers in \emph{RangeShiftR} are naïve and have no goal, it may be applied only in the ‘negative’
sense of moving away from the natal location (\code{GoalType=2}), i.e. as a dispersal bias, which is subject to a decay in strength as a
function of the total number of steps taken (set the decay rate \code{AlphaDB} and inflection point \code{BetaDB}).
This enables a dispersal path to follow a straighter trajectory initially, and later become more responsive to perceived landscape costs.

The product of the reciprocals of effective cost, \code{DP} and dispersal bias, scaled to sum to one,
give the probabilities that the individual will move to each neighbouring cell. All the dispersing individuals move simultaneously, i.e. at
each time-step they all make one move. In the case of patch-based models, the individual is forced to leave the natal patch by increasing its
\code{DP} ten-fold until it has taken a number of steps (\eqn{=2*}\code{PR}) outside the natal patch.

When an individual arrives in a non-natal patch and decides not to settle there (as a result of a density-dependent or mate-finding settlement
rule; see \code{\link[RangeShiftR]{Settlement}}), then there is the option that its path is straightened (\code{StraightenPath=TRUE}). This means
that it leaves the patch as soon as possible in order to search for another patch. This is achieved by increasing its \code{DP} ten-fold. However, in
certain types of model, e.g. when arbitrary contiguous patches have been defined for what is basically a continuous population, this can lead
to the path always being straightened, as an individual enters a new patch as soon as it has left the one it has rejected. In such cases, it is
advisable to disable the feature (\code{StraightenPath=FALSE}), although care must be taken that individuals do not become trapped in patches
surrounded by very high cost matrix.

When inter-individual variability is activated (set \code{IndVar=TRUE}), the four SMS movement traits \code{DP}, as well as - if dispersal goal bias
is enabled - \code{GB}, \code{AlphaDB} and \code{BetaDB} can evolve.
For each trait the population initial mean and standard deviations must be set (instead of the constant value), as well as its scaling factor
(see \code{\link[RangeShiftR]{Genetics}}).\cr

As currently implemented, there is no sex- or stage- dependence of SMS traits.

The production of the output ‘dispersal heat maps’, which show the total number of times each cell is visited by dispersing individuals within a
single replicate simulation, can be enabled with the option \code{SMSHeatMap} in \code{\link[RangeShiftR]{Simulation}}.

\emph{Costs layer} \cr
Critical for the outcomes of SMS are the relative costs assigned to the different habitats (as it is also the case for
the LCP approach). Habitat costs or resistance to movement can be set manually for each habitat code or imported as a raster map, which
allows for costs to be a function of multiple variables instead of a simple value associated to the habitat type.
In the latter case, the file names are provided in the \code{\link[RangeShiftR]{ImportedLandscape}} module.
The specified map has to match the landscape raster in extent, coordinates and resolution, and each cell contains a cost value, with the minimal possible cost being \eqn{1}.
Importing a cost layer is the only option when the landscape comprises habitat coverage or quality.

\emph{Mortality} \cr
There are two main sources of mortality: First, dispersal mortality can arise as a result of individuals failing to reach suitable habitat. Some individuals may fail
to find suitable habitat before they use up their maximum number of movement steps (\code{MaxSteps} in \code{\link[RangeShiftR]{Settlement}}).
In this first case, dispersal mortality clearly depends upon the proportion of suitable habitat in the landscape and will increase as the
availability of habitat declines.\cr
A second source of dispersal mortality can be specified by the user in form of a per-step probability of mortality (\code{StepMort}.
This can be useful for representing mortality risks that increase with distance or time spent travelling \insertCite{bonte2012costs}{RangeShiftR}.
Additionally, it is possible that the per-step mortality varies according to the nature of the local environment by providing the \code{CostMap}.

Note that the total dispersal mortality
experienced will be the sum of the mortalities due to the two sources identified above and, in parameterising the model, it will be important
to recognize this such that dispersal mortality is not double-accounted.
}
\references{
\insertAllCited{}
}
\author{
Anne-Kathleen Malchow
}
