% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_handling.R
\name{SMSpathLengths}
\alias{SMSpathLengths}
\title{Get the distribution of SMS Path Lengths}
\usage{
SMSpathLengths(s, dirpath, ...)
}
\arguments{
\item{s}{RSmaster parameter object}

\item{dirpath}{RS directory path}
}
\value{
a data.frame that contains the mean (over replicates) number of SNS paths of a given length (rows) per simulated map (columns)
}
\description{
Reads the RangeShiftR output files 'MovePaths' (if they were generated) to get the distribution of lengths of SMS
paths taken in all recorded years.
}
