% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DispersalParams.R
\docType{class}
\name{Transfer}
\alias{Transfer}
\alias{TransferParams}
\title{Set a Transfer method}
\description{
Transfer (or transience) is the second phase of dispersal. It consists of the movement of an individual departing from its natal patch towards
a potential new patch, ending with settlement or mortality. This movement can be modelled by one of three alternative processes:\cr
- Dispersal kernel: use \code{\link[RangeShiftR]{DispersalKernel}}\cr
- Stochastic movement simulator (SMS): use \code{\link[RangeShiftR]{SMS}}\cr
- Correlated random walk (CRW): use \code{\link[RangeShiftR]{CorrRW}}
}
\details{
The choice between the two main methods to model the transfer phase, i.e. phenomenological dispersal kernels or mechanistic movement processes
(SMS and CRW) depends on the information available for a given species and on the level of detail that is considered important to represent in
the model (which will depend on its aim and the scale).
\cr
Dispersal is often a costly process for an organism \insertCite{bonte2012costs}{RangeShiftR} and, in some cases, a dispersing individual may
suffer mortality. Obtaining a sensible
representation of dispersal requires that these mortality costs are described appropriately. The total dispersal mortality experienced will
be the sum of two main sources of mortality: First, as the result of individuals failing to reach suitable habitat, and second, as the result
of increased energetic, time or attritional costs that longer-distance dispersers will experience.
For more details, refer also to the respective Details sections of the different transfer methods.
In parameterising the model, it will be important to recognise this such that dispersal mortality is not double-accounted.
}
\references{
\insertAllCited{}
}
\author{
Anne-Kathleen Malchow
}
