% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProbs.R
\name{plotProbs}
\alias{plotProbs}
\title{Plot parameterised probabilities}
\usage{
plotProbs(x, ...)
}
\arguments{
\item{x}{a RangeShiftR parameter object}

\item{...}{various options, depend on the given parameter module type}

\item{xmax, ymax}{upper plotting limits (lower limits are fixed to \eqn{0})}

\item{stage, sex}{stage(s) and sexe(s) to plot, default: all}
}
\description{
Visualise the dependencies or distributions of probabilities of the various processes that are defined by the different parameter modules,
like e.g. fecundity or mortality.
}
\details{
Available methods and their options:
\itemize{
  \item \code{\link[RangeShiftR]{Emigration}}: plot emigration probability
  \item \code{\link[RangeShiftR]{DispersalKernel}}:
  \itemize{
    \item \code{mortality=FALSE} - plot dispersal distance probability density  (default)
    \item \code{mortality= TRUE} - plot mortality probability
  }
  If a mixed kernel was defined (i.e. \code{DoubleKernel=TRUE}), plot the resulting dispersal probability by...
  \itemize{
    \item \code{combinekernels=FALSE} - ...plotting both kernels separately (default)
    \item \code{combinekernels= TRUE} - ...combining both kernels, i.e. \ifelse{html}{ \out{p(d; &delta;<sub>1</sub>,&delta;<sub>2</sub>) = p<sub>I</sub> p(d;&delta;<sub>1</sub>) + (1-p<sub>I</sub>) p(d;&delta;<sub>1</sub>) } }{\deqn{ p(d; δ_1,δ_2) = p_I p(d;δ_1) + (1-p_I) p(d;δ_2)} }
  }
  \item \code{\link[RangeShiftR]{StageStructure}}: plot fecundity as well as survival and development probabilities
}
}
