% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeShiftR.R
\docType{package}
\name{RangeShiftR-package}
\alias{RangeShiftR}
\alias{RangeShiftR-package}
\title{RangeShiftR: An R package for individual-based simulation of spatial eco-evolutionary dynamics and species' responses to environmental changes}
\description{
RangeShiftR provides individual-based simulations of spatial eco-evolutionary dynamics. It is based on the C++ software RangeShifter, making it flexible and fast. RangeShiftR models the processes of demography, dispersal and evolution in an inter-dependent way, offering substantial complexity in the corresponding modelling choices. It is entirely open-source and aims to facilitate the application of individual-based and mechanistic modelling to eco-evolutionary questions.
}
\details{
\code{RangeShiftR} provides flexible and fast simulations of spatial eco-evolutionary dynamics.
It implements the individual-based simulation software \emph{RangeShifter},
\insertCite{bocedi2014,bocedi2020}{RangeShiftR},
making it available for all machines that can build and run \emph{R}.

\code{RangeShiftR} models the processes of demography, dispersal and evolution in an
inter-dependent way, offering substantial complexity in the corresponding modelling choices.
It is entirely open-source and aims to facilitate the application
of individual-based and mechanistic modelling to eco-evolutionary questions.

The \emph{RangeShifter} project is presented on our website
\url{https://rangeshifter.github.io/}.

To get started with the package, please find an overview and a range of tutorials here:
\url{https://rangeshifter.github.io/RangeshiftR-tutorials/}.
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rangeshifter.github.io/RangeshiftR-tutorials/}
}

}
\author{
\strong{Maintainer}: Anne-Kathleen Malchow \email{rangeshiftr@uni-potsdam.de}

Authors:
\itemize{
  \item Greta Bocedi
  \item Stephen C.F. Palmer
  \item Justin M.J. Travis
  \item Damaris Zurell
}

}
\keyword{internal}
